+++ project/src/pages/login-screen/login-screen.tsx
@@ -4,7 +4,7 @@ import { Header } from '../../components/header/header';
 import { AppRoute } from '../../consts/app-route';
 import { AuthorizationStatus } from '../../consts/authorization-status';
 import { useAppDispatch, useAppSelector } from '../../hooks';
-import { loginAction } from '../../store/api-actions';
+import { loginAction } from '../../services/store/api-actions';
 import { AuthData } from '../../types/auth-data';
 
 export const LoginScreen: React.FunctionComponent = () => {
