+++ project/src/services/process-error-handle.ts
@@ -1,6 +1,6 @@
-import { store } from '../store';
-import { setError } from '../store/action';
-import { clearErrorAction } from '../store/api-actions';
+import { store } from './store';
+import { setError } from './store/action';
+import { clearErrorAction } from './store/api-actions';
 
 export const processErrorHandle = (message: string): void => {
   store.dispatch(setError(message));
