+++ project/src/services/store/index.ts
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { reducer } from './reducer';
-import { createAPI } from '../services/api';
+import { createAPI } from '../api';
 import { redirect } from './middlewares/redirect';
 
 export const api = createAPI();
